+++ modules/items/item-sheet-ffg-v2.js
@@ -5,7 +5,7 @@ export class ItemSheetFFGV2 extends ItemSheetFFG {
   static get defaultOptions() {
     return foundry.utils.mergeObject(super.defaultOptions, {
       classes: ["starwarsffg", "sheet", "item", "v2"],
-      tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "description" }],
+      tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "attributes" }],
       scrollY: [".sheet-body", ".tab"],
     });
   }